backend/requirements.txt
fastapi
uvicorn
python-multipart
g2p_en
Levenshtein
librosa
numpy
scipy
soundfile
edge-tts
openai-whisper
